Find existing quadlets more precisely
During validation of a new quadlet the existing quadlet must
be copied to an ephemeral directory.

This failed with:

```
cp: -r not specified; omitting directory '/etc/containers/systemd/users'
```

The extra directory should not be copied. The actually files with
be `my.pod`, `my.container`, ...

Particularly on Ubuntu and Debian the directory `/etc/containers/systemd/users`
exists to store system provided user quadlets.